+++ NEWS.md
@@ -1,5 +1,6 @@
 ## 2023-XX-XX v3.1.0-SNAPSHOT
 * [MODDATAIMP-1007](https://issues.folio.org/browse/MODDATAIMP-1007) Make endpoint as async.
+* [MODDATAIMP-1015](https://issues.folio.org/browse/MODDATAIMP-1015) Upgrade data-import to RMB v35.2.0 and Vertx 4.5.4
 
 ## 2023-XX-XX v3.1.0
 * [MODDATAIMP-1003](https://folio-org.atlassian.net/browse/MODDATAIMP-1003)